Removes active record dependencies
In order to utilize deleted_at with Rails 5+ we need to remove the
contraints around active record. This led to a mismatch with the way
deleted at was querying postgres to know if tables were installed. The
select statements were altered to make a cleaner predicate method in
determining if the views had been installed or not.